fix(remove componentsview): remove test component and move custom-component to componentView.vue
